requirements.txt
# Raspberry Pi hardware
RPi.GPIO>=0.7.1
picamera2>=0.3.12

# Web framework
flask>=3.0
flask-cors>=4.0

# Face recognition
face-recognition>=1.3.0
numpy>=1.24

# Push notifications
firebase-admin>=6.0

# Image handling
Pillow>=10.0
